+++ example/lib/app/app_module.dart
@@ -1,27 +1,19 @@
 import 'package:flutter_modular/flutter_modular.dart';
-import 'package:flutter/material.dart';
-import 'package:i18n_remote_config/i18n_remote_config.dart';
 
 import 'app_controller.dart';
-import 'app_widget.dart';
+
 import 'modules/home/home_module.dart';
 import 'splash/splash_module.dart';
 
-class AppModule extends MainModule {
+class AppModule extends Module {
   @override
   List<Bind> get binds => [
         Bind((i) => AppController()),
-        Bind((i) => I18nLocalizations()),
       ];
 
   @override
-  List<ModularRouter> get routers => [
-        ModularRouter('/', module: SplashModule()),
-        ModularRouter('/home', module: HomeModule()),
+  List<ModularRoute> get routes => [
+        ModuleRoute('/', module: SplashModule()),
+        ModuleRoute('/home', module: HomeModule()),
       ];
-
-  @override
-  Widget get bootstrap => AppWidget();
-
-  static Inject get to => Inject<AppModule>.of();
 }