+++ example/lib/app/app_widget.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_modular/flutter_modular.dart';
+import 'package:i18n_remote_config/i18n_remote_config.dart';
 
 class AppWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      navigatorKey: Modular.navigatorKey,
       title: 'Flutter Slidy',
       theme: ThemeData(
         primarySwatch: Colors.blue,
@@ -18,7 +18,6 @@ class AppWidget extends StatelessWidget {
       ],
       supportedLocales: [Locale("pt", "BR")],
       initialRoute: '/',
-      onGenerateRoute: Modular.generateRoute,
-    );
+    ).modular();
   }
 }