+++ example/lib/app/modules/home/home_module.dart
@@ -2,16 +2,14 @@ import 'home_controller.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'home_page.dart';
 
-class HomeModule extends ChildModule {
+class HomeModule extends Module {
   @override
   List<Bind> get binds => [
         Bind((i) => HomeController()),
       ];
 
   @override
-  List<ModularRouter> get routers => [
-        ModularRouter('/', child: (_, args) => HomePage()),
+  List<ModularRoute> get routes => [
+        ChildRoute('/', child: (_, args) => HomePage()),
       ];
-
-  static Inject get to => Inject<HomeModule>.of();
 }