+++ example/lib/main.dart
@@ -1,5 +1,9 @@
+import 'package:example/app/app_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:example/app/app_module.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 
-void main() => runApp(ModularApp(module: AppModule()));
+void main() => runApp(ModularApp(
+      module: AppModule(),
+      child: AppWidget(),
+    ));