+++ example/test/app/modules/home/home_controller_test.dart
@@ -1,4 +1,5 @@
-import 'package:flutter_modular/flutter_modular_test.dart';
+import 'package:flutter_modular/flutter_modular.dart';
+import 'package:flutter_modular_test/flutter_modular_test.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:example/app/modules/home/home_controller.dart';
@@ -9,7 +10,7 @@ void main() {
   late HomeController home;
 
   setUp(() {
-    home = HomeModule.to.get<HomeController>();
+    home = Modular.get<HomeController>();
   });
 
   group('HomeController Test', () {