+++ example/test/app/modules/home/home_page_test.dart
@@ -1,5 +1,5 @@
+import 'package:flutter_modular_test/flutter_modular_test.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_modular/flutter_modular_test.dart';
 
 import 'package:example/app/modules/home/home_page.dart';