+++ lib/src/i18n_string_ext.dart
@@ -1,5 +1,3 @@
-// ignore: import_of_legacy_library_into_null_safe
-import 'package:flutter_modular/flutter_modular.dart';
 import 'package:i18n_remote_config/i18n_remote_config.dart';
 
 class I18nString {}
@@ -7,7 +5,7 @@ class I18nString {}
 extension I18nStringExt on String {
   String i18n({List<String>? params}) {
     final key = this;
-    final localization = Modular.get<I18nLocalizations>();
+    final localization = I18nLocalizations.instance;
     var value = localization.getValue(key);
     if (params != null) {
       for (var i = 0; i < params.length; i++) {