+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+## [0.0.4-nullsafety+2] - TODO: Removed packages.
+
+- Removed Flutter Module
+- Create Singleton for comunication
+
 ## [0.0.4-nullsafety+1] - TODO: Migration Nullsafety.
 
 - Update dependencies
+
 ## [0.0.4-nullsafety] - TODO: Migration Nullsafety.
 
 ## [0.0.3+1] - TODO: Support release firebase.